requirements.txt
streamlit==1.32.0
amadeus==12.0.0
python-dotenv==1.1.1
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.4.0